Added a fix for date conversion
Date conversions were failing in IE 11 due to the fact the the string
replace function would only replace a single instance of a character
instead of all of them.  This has been changed to a regular expression
that forces all of the characters to convert.  This wasn't an issue in
Chrome for whatever reason.